% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFP.R
\docType{data}
\name{BFP}
\alias{BFP}
\title{Business failure prediction demonstration data set}
\description{
Business failure prediction demonstration data set. Contains financial ratios and firmographics as independent variables for 522 anonymized European companies. The Class column indicates failure (class 1) or survival (class 0) over a 1-year period.
}
\references{
De Bock, K.W., Lessmann, S. And Coussement, K., Cost-sensitive business failure prediction
when misclassification costs are uncertain: A heterogeneous ensemble selection approach,
European Journal of Operational Research (2020), doi: 10.1016/j.ejor.2020.01.052.
}
\author{
Koen W. De Bock, \email{kdebock@audencia.com}
}
