% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFP.R
\docType{data}
\name{BFP}
\alias{BFP}
\title{Business failure prediction demonstration data set}
\description{
Business failure prediction demonstration data set. Contains financial ratios and firmographics as independent variables for 522 anonymized European companies. The Class column indicates failure (class 1) or survival (class 0) over a 1-year period.
}
\references{
De Bock, K.W., Lessmann, S. And Coussement, K., 2014, Multicriteria optimization for cost-sensitive ensemble
selection in business failure prediction, Proc. 20th Conference of the International Federation of Operational Research Societies (IFORS 2014), Barcelona, Spain.
}
\author{
Koen W. De Bock, \email{kdebock@audencia.com}
}
